proposal for 2nd round of postprocessing

+++ cmgPostProcessing/cmgPostProcessing.py
@@ -1,7 +1,6 @@
 import ROOT
 import sys, os, copy, random, subprocess, datetime
 from array import array
-#from StopsDilepton.tool.cmgObjectSelection import cmgLooseLepIndices, splitIndList, get_cmg_jets_fromStruct, splitListOfObjects, cmgTightMuID, cmgTightEleID
 from StopsDilepton.tools.convertHelpers import compileClass, readVar, printHeader, typeStr, createClassString
 
 from math import *
@@ -9,22 +8,24 @@ from StopsDilepton.tools.mt2Calculator import mt2Calculator
 mt2Calc = mt2Calculator()
 from StopsDilepton.tools.mtautau import mtautau as mtautau_
 from StopsDilepton.tools.helpers import getChain, getChunks, getObjDict, getEList, getVarValue
-from StopsDilepton.tools.objectSelection import getLeptons, looseMuID, looseEleID, getJets
+from StopsDilepton.tools.objectSelection import getLeptons, getMuons, getElectrons, getGoodMuons, getGoodElectrons, getGoodLeptons, getJets, getGoodBJets, getGoodJets, isBJet 
+
 from StopsDilepton.tools.localInfo import *
 
 ROOT.gSystem.Load("libFWCoreFWLite.so")
 ROOT.AutoLibraryLoader.enable()
 
 from StopsDilepton.samples.xsec import xsec
+from StopsDilepton.samples.cmgTuples_Phys14_signals import *
 from StopsDilepton.samples.cmgTuples_Data50ns_1l import *
 from StopsDilepton.samples.cmgTuples_Spring15_25ns import *
 from StopsDilepton.samples.cmgTuples_Spring15_50ns import *
 
-target_lumi = 4000 #pb-1 Which lumi to normalize to
+target_lumi = 1000 #pb-1 Which lumi to normalize to
 
-defSampleStr = "TBar_tWch_50ns"  #Which samples to run for by default (will be overritten by --samples option)
+defSampleStr = "MuonEG_Run2015B_PromptReco"  #Which samples to run for by default (will be overritten by --samples option)
 
-subDir = "/data/rschoefbeck/cmgTuples/postProcessed_Spring15" #Output directory -> The first path should go to localInfo (e.g. 'dataPath' or something)
+subDir = "/data/rschoefbeck/cmgTuples/postProcessed_Spring15_pass2" #Output directory -> The first path should go to localInfo (e.g. 'dataPath' or something)
 
 #branches to be kept for data and MC
 branchKeepStrings_DATAMC = ["run", "lumi", "evt", "isData", "rho", "nVert", 
@@ -59,6 +60,7 @@ parser.add_option("--inputTreeName", dest="inputTreeName", default="treeProducer
 parser.add_option("--targetDir", dest="targetDir", default=subDir, type="string", action="store", help="target directory.")
 parser.add_option("--skim", dest="skim", default="dilep", type="string", action="store", help="any skim condition?")
 parser.add_option("--small", dest="small", default = False, action="store_true", help="Just do a small subset.")
+parser.add_option("--overwrite", dest="overwrite", default = False, action="store_true", help="Overwrite?")
 
 (options, args) = parser.parse_args()
 assert options.skim in ['none', 'dilep'], "Unknown skim: %s"%options.skim
@@ -92,36 +94,45 @@ def getTreeFromChunk(c, skimCond, iSplit, nSplit):
    
 exec('allSamples=['+options.allSamples+']')
 for isample, sample in enumerate(allSamples):
-  #chunks, sumWeight = getChunks(sample, options.inputTreeName)
+  outDir = options.targetDir+'/'+"/".join([options.skim, sample['name']])
+
   maxN = 1 if options.small else -1
   chunks, sumWeight = getChunks(sample, maxN=maxN)
-  #chunks, nTotEvents = getChunksFromDPM(sample, options.inputTreeName)
-#  print "Chunks:" , chunks 
-  outDir = options.targetDir+'/'+"/".join([options.skim, sample['name']])
+  
+  if os.path.exists(outDir) and os.listdir(outDir) != [] and not options.overwrite:
+    print "Found non-empty directory: %s -> skipping!"%outDir
+    continue
   tmpDir = outDir+'/tmp/'
-  os.system('mkdir -p ' + outDir) 
+  os.system('mkdir -p '+outDir) 
   os.system('mkdir -p '+tmpDir)
   os.system('rm -rf '+tmpDir+'/*')
-  
+
   if sample['isData']: 
     lumiScaleFactor=1
     branchKeepStrings = branchKeepStrings_DATAMC + branchKeepStrings_DATA 
+    jetMCInfo = []
   else:
     lumiScaleFactor = xsec[sample['dbsName']]*target_lumi/float(sumWeight)
     branchKeepStrings = branchKeepStrings_DATAMC + branchKeepStrings_MC
+    jetMCInfo = ['mcMatchFlav/I', 'partonId/I']
 
   readVariables = ['met_pt/F', 'met_phi/F']
   newVariables = ['weight/F']
   aliases = [ "met:met_pt", "metPhi:met_phi"]
-
   readVectors = [\
     {'prefix':'LepGood',  'nMax':8, 'vars':['pt/F', 'eta/F', 'phi/F', 'pdgId/I', 'charge/I', 'relIso03/F', 'tightId/I', 'miniRelIso/F','mass/F','sip3d/F','mediumMuonId/I', 'mvaIdPhys14/F','lostHits/I', 'convVeto/I', 'dxy/F', 'dz/F']},
-    {'prefix':'Jet',  'nMax':100, 'vars':['pt/F', 'eta/F', 'phi/F', 'id/I','btagCSV/F', 'btagCMVA/F']}, #, 'mcMatchFlav/I', 'partonId/I']},
+    {'prefix':'Jet',  'nMax':100, 'vars':['pt/F', 'eta/F', 'phi/F', 'id/I','btagCSV/F', 'btagCMVA/F'] + jetMCInfo},
   ]
   if not sample['isData']: 
     aliases.extend(['genMet:met_genPt', 'genMetPhi:met_genPhi'])
-#  if options.skim.lower() in ['dilep']:
-#    newVariables.extend( ['mll/F'] )
+  if options.skim.lower() in ['dilep']:
+    newVariables.extend( ['nGoodMuons/I', 'nGoodElectrons/I' ] )
+    newVariables.extend( ['dl_pt/F', 'dl_eta/F', 'dl_phi/F', 'dl_mass/F' ] )
+    newVariables.extend( ['dl_mt2ll/F', 'dl_mt2bb/F', 'dl_mt2blbl/F', 'dl_mtautau/F', 'dl_alpha0/F',  'dl_alpha1/F' ] )
+    newVariables.extend( ['l1_pt/F', 'l1_eta/F', 'l1_phi/F', 'l1_mass/F', 'l1_pdgId/I', 'l1_index/I' ] )
+    newVariables.extend( ['l2_pt/F', 'l2_eta/F', 'l2_phi/F', 'l2_mass/F', 'l2_pdgId/I', 'l2_index/I' ] )
+    newVariables.extend( ['isEE/I', 'isMuMu/I', 'isEMu/I', 'isOS/I' ] )
+
   newVars = [readVar(v, allowRenaming=False, isWritten = True, isRead=False) for v in newVariables]
   
   readVars = [readVar(v, allowRenaming=False, isWritten=False, isRead=True) for v in readVariables]
@@ -176,43 +187,62 @@ for isample, sample in enumerate(allSamples):
         t.GetEntry(i)
         genWeight = 1 if sample['isData'] else t.GetLeaf('genWeight').GetValue()
         s.weight = lumiScaleFactor*genWeight
-          #print "reweighted:" , s.weight
+        if options.skim.lower()=='dilep':
+          leptons = getGoodLeptons(r)
+          s.nGoodMuons      = len(filter( lambda l:abs(l['pdgId'])==13, leptons))
+          s.nGoodElectrons  = len(filter( lambda l:abs(l['pdgId'])==11, leptons))
+#          print "Leptons", leptons 
+          if len(leptons)>=2:# and leptons[0]['pdgId']*leptons[1]['pdgId']<0 and abs(leptons[0]['pdgId'])==abs(leptons[1]['pdgId']): #OSSF choice
+            mt2Calc.reset()
+            s.l1_pt  = leptons[0]['pt'] 
+            s.l1_eta = leptons[0]['eta']
+            s.l1_phi = leptons[0]['phi']
+            s.l1_mass   = leptons[0]['mass']
+            s.l1_pdgId  = leptons[0]['pdgId']
+            s.l1_index  = leptons[0]['index']
+            s.l2_pt  = leptons[1]['pt'] 
+            s.l2_eta = leptons[1]['eta']
+            s.l2_phi = leptons[1]['phi']
+            s.l2_mass   = leptons[1]['mass']
+            s.l2_pdgId  = leptons[1]['pdgId']
+            s.l2_index  = leptons[1]['index']
 
-#        if options.skim=='dilep':
-#          mt2Calc.reset()
-#          #Leptons 
-#          leptons = filter(lambda l: looseMuID(l) or looseEleID(l), getLeptons(r))
-#          
-#          if len(leptons)>=2:# and leptons[0]['pdgId']*leptons[1]['pdgId']<0 and abs(leptons[0]['pdgId'])==abs(leptons[1]['pdgId']): #OSSF choice
-#            l0pt, l0eta, l0phi = leptons[0]['pt'],  leptons[0]['eta'],  leptons[0]['phi']
-#            l1pt, l1eta, l1phi = leptons[1]['pt'],  leptons[1]['eta'],  leptons[1]['phi']
-#            mll = sqrt(2.*l0pt*l1pt*(cosh(l0eta-l1eta)-cos(l0phi-l1phi)))
-#            jets = filter(lambda j:j['pt']>30 and abs(j['eta'])<2.4 and j['id'], getJets(r))
-#            bjets = filter(lambda j:j['btagCSV']>0.890, jets)
-#            if len(bjets)==2:
-#              mt2Calc.setMet(met,metPhi)
-#              mt2Calc.setLeptons(l0pt, l0eta, l0phi, l1pt, l1eta, l1phi)
-#              mt2ll = mt2Calc.mt2ll()
-#    #          if mt2ll>120:
-#              if True:
-#                plots['mt2ll']['histo'][s["name"]].Fill(mt2ll, weight)
-#                ht = sum([j['pt'] for j in jets])
-#                plots['kinMetSig']['histo'][s["name"]].Fill(met/sqrt(ht), weight)
-#                mt2Calc.setBJets(bjets[0]['pt'], bjets[0]['eta'], bjets[0]['phi'], bjets[1]['pt'], bjets[1]['eta'], bjets[1]['phi'])
-#                mt2bb   = mt2Calc.mt2bb()
-#                mt2blbl = mt2Calc.mt2blbl()
-#                plots['mt2bb']['histo'][s["name"]].Fill(mt2bb, weight)
-#                plots['mt2blbl']['histo'][s["name"]].Fill(mt2blbl, weight)
-#                mtautau, alpha_0, alpha_1 = mtautau_(met,metPhi, l0pt, l0eta, l0phi, l1pt, l1eta, l1phi, retAll=True)
-#                plots['mtautau']['histo'][s["name"]].Fill(mtautau, weight)
-#                plots['mtautau_zoomed']['histo'][s["name"]].Fill(mtautau, weight)
+            l_pdgs = [abs(leptons[0]['pdgId']), abs(leptons[1]['pdgId'])]
+            l_pdgs.sort()
+            s.isMuMu = l_pdgs==[13,13] 
+            s.isEE = l_pdgs==[11,11] 
+            s.isEMu = l_pdgs==[11,13] 
+            s.isOS = s.l1_pdgId*s.l2_pdgId<0
+
+            l1 = ROOT.TLorentzVector()
+            l1.SetPtEtaPhiM(leptons[0]['pt'], leptons[0]['eta'], leptons[0]['phi'], 0 )
+            l2 = ROOT.TLorentzVector()
+            l2.SetPtEtaPhiM(leptons[1]['pt'], leptons[1]['eta'], leptons[1]['phi'], 0 )
+            dl = l1+l2
+            s.dl_pt  = dl.Pt()
+            s.dl_eta = dl.Eta()
+            s.dl_phi = dl.Phi()
+            s.dl_mass   = dl.M() 
+            mt2Calc.setMet(r.met_pt,r.met_phi)
+            mt2Calc.setLeptons(s.l1_pt, s.l1_eta, s.l1_phi, s.l2_pt, s.l2_eta, s.l2_phi)
+            s.dl_mt2ll = mt2Calc.mt2ll()
+            s.dl_mtautau, s.dl_alpha0, s.dl_alpha1 = mtautau_(r.met_pt,r.met_phi, s.l1_pt, s.l1_eta, s.l1_phi, s.l2_pt, s.l2_eta, s.l2_phi, retAll=True)
+
+            jets = getGoodJets(r)
+            if len(jets)>=2:
+              bJets = filter(lambda j:isBJet(j), jets)
+              nonBJets = filter(lambda j:not isBJet(j), jets)
+              bj0, bj1 = (bJets+nonBJets)[:2]
+              mt2Calc.setBJets(bj0['pt'], bj0['eta'], bj0['phi'], bj1['pt'], bj1['eta'], bj1['phi'])
+              s.dl_mt2bb   = mt2Calc.mt2bb()
+              s.dl_mt2blbl = mt2Calc.mt2blbl()
+#              print len(bJets), len(nonBJets), s.dl_mt2bb, s.dl_mt2blbl
 
         for v in newVars:
           v['branch'].Fill()
       newFileName = sample['name']+'_'+chunk['name']+'_'+str(iSplit)+'.root'
       filesForHadd.append(newFileName)
-      if True or  not options.small:
-      #if options.small:
+      if not options.small:
         f = ROOT.TFile(tmpDir+'/'+newFileName, 'recreate')
         t.SetBranchStatus("*",0)
         for b in branchKeepStrings + [v['stage2Name'] for v in newVars] +  [v.split(':')[1] for v in aliases]:
@@ -229,6 +259,7 @@ for isample, sample in enumerate(allSamples):
         del v['branch']
 
   print "Event loop end"
+
   if not options.small: 
     size=0
     counter=0
@@ -236,7 +267,7 @@ for isample, sample in enumerate(allSamples):
     for f in filesForHadd:
       size+=os.path.getsize(tmpDir+'/'+f)
       files.append(f)
-      if size>(0.5*(10**9)) or f==filesForHadd[-1]:
+      if size>(0.5*(10**9)) or f==filesForHadd[-1] or len(files)>300:
         ofile = outDir+'/'+sample['name']+'_'+str(counter)+'.root'
         print "Running hadd on", tmpDir, files
         os.system('cd '+tmpDir+';hadd -f '+ofile+' '+' '.join(files))
